requirements.txt
asttokens==3.0.0
colorama==0.4.6
et_xmlfile==2.0.0
executing==2.2.0
icecream==2.1.4
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
PuLP==3.2.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
xlrd==2.0.1
